Revert "Use safer comparison for TOTP tokens"
This reverts commit d69435b4d91769a8f84ecd0e32ebda607fc9de85 and introduces a
CHANGELOG entry about why.

A summary of the bug is as follows: the previous version did introduce a timing
differential on the length of the TOTP code. This was discussed in the PR (#11),
but an underlying issue was missed in that discussion. The rationale for the
short-circuit branch on the length of the TOTP code differential didn't matter
is twofold:

  a) without it, you have a truncation vulnerability; e.g. the attacker
     submits `1` and any TOTP code starting with `1` is accepted

  b) the length of the TOTP code is public anyway.

That's true, but misses another point. By random chance, the legitimate TOTP
code will sometimes be a number small enough that it doesn't require
all (usually 6) digits to represent. 10% of the time the first digit will be
zero, 1% of the time the first two will be, and so on. The string conversion
will result in a shorter code string, because it doesn't know to pad with
zeroes. If the attacker submits full-length TOTP codes, the fast short-circuit
tells them the real code is shorter. The attacker can then submit shorter and
shorter codes until the timing differential tells them they've found the right
length. Finally, they only need to be patient enough to hit a 30-second window
where the remaining nonzero codes are enumerable.

Let's put some numbers on how likely that is. The probability of any particular
code having N prefix zeroes is 10^(-N). The probability of seeing any such code
over M codes is 1-(1-10^(-N))^M. A day is 86400 seconds; given the standard
refresh of a TOTP of 30s, that’s 2880 values per day. For N=3 prefix zeroes and
M=2880 (the probability of seeing at least one code with 3 prefix zeroes over 24
hours) is about 94%. For N=4, M=2880, 25%. For N=5, M=2880, 3%.

There may be a way to rehabilitate this approach but in the mean while reverting
to the probably-not-exploitable-but-dependent-on-the-caller version is.

Thanks to @cursork and @sw1nn for raising this! Sorry it took so long to see it.